# Conflicts:
#	yukon-client/multispeak/src/com/cannontech/multispeak/client/core/v4/CBClient.java
#	yukon-client/multispeak/src/com/cannontech/multispeak/client/core/v4/ICBClient.java
#	yukon-client/multispeak/src/com/cannontech/multispeak/client/v4/MultispeakFuncs.java
#	yukon-client/multispeak/src/com/cannontech/multispeak/dao/impl/v4/MspObjectDaoImpl.java
#	yukon-client/multispeak/src/com/cannontech/multispeak/dao/v4/MspObjectDao.java
#	yukon-client/yukon-web/src/com/cannontech/web/multispeak/MspHandler.java

+++ yukon-client/yukon-web/src/com/cannontech/web/capcontrol/FeederController.java
@@ -245,6 +245,11 @@ public class FeederController {
                     flash.setError(new YukonMessageSourceResolvable(feederKey + ".delete.voltagePoint.error", feeder.getName()));
                     return "redirect:/capcontrol/feeders/" + id;
                 }
+                boolean isFeederAssignedToRegulatorPoint = feederService.isFeederAssignedToRegulatorPointForZone(id);
+                if (isFeederAssignedToRegulatorPoint) {
+                    flash.setError(new YukonMessageSourceResolvable(feederKey + ".delete.regulatorPoint.error", feeder.getName()));
+                    return "redirect:/capcontrol/feeders/" + id;
+                }
                 Integer parentId = feederDao.getParentSubBusID(id);
                 feederService.delete(id);
                 flash.setConfirm(new YukonMessageSourceResolvable(feederKey + ".delete.success", feeder.getName()));
